+++ src/main/java/sejong/smartnotice/service/SupporterService.java
@@ -13,7 +13,7 @@ import sejong.smartnotice.domain.member.Account;
 import sejong.smartnotice.domain.member.Supporter;
 import sejong.smartnotice.domain.member.User;
 import sejong.smartnotice.dto.SupporterModifyDTO;
-import sejong.smartnotice.dto.SupporterRegisterDTO;
+import sejong.smartnotice.dto.request.SupporterRegisterDTO;
 import sejong.smartnotice.repository.SupporterRepository;
 
 import java.util.List;
